% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{api1_url}
\alias{api1_url}
\title{Get full Genesys API v1 URL for a specific path}
\usage{
api1_url(path)
}
\arguments{
\item{path}{relative path of the API v1 endpoint (e.g. \code{/me})}
}
\value{
Absolute URL to an API call
}
\description{
Get full Genesys API v1 URL for a specific path
}
\examples{
 api1_url("/me")
}
