% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{api_url}
\alias{api_url}
\title{Get full Genesys API URL for a specific path}
\usage{
api_url(path)
}
\arguments{
\item{path}{relative path of the API endpoint (e.g. \code{/me})}
}
\value{
Absolute URL to an API call
}
\description{
Get full Genesys API URL for a specific path
}
\examples{
 api_url("/me")
}
