% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{.list_accessions_page}
\alias{.list_accessions_page}
\title{List accession passport data (paginated)}
\usage{
.list_accessions_page(
  filters = list(),
  page = 0,
  size = 1000,
  fields = NULL,
  selector = NULL
)
}
\value{
table
}
\description{
List accession passport data (paginated)
}
\keyword{internal}
