% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\docType{data}
\name{.LISTENER}
\alias{.LISTENER}
\title{Reference to the server}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.LISTENER
}
\description{
Reference to the server
}
\keyword{internal}
