% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{download_mcpd}
\alias{download_mcpd}
\title{Download all passport data for one genebank in Excel format and save it to disk}
\usage{
download_mcpd(instituteCode, file = NULL)
}
\arguments{
\item{instituteCode}{FAO WIEWS institute code}

\item{file}{Target file name. Defaults to Genesys-provided file name in the current working directory.}
}
\value{
HTTP response data
}
\description{
Download all passport data for one genebank in Excel format and save it to disk
}
\examples{
\dontrun{
  # Download MCPD passport data for NGA039
  excelData <- download_mcpd("NGA039")
}

}
