% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mcpd_filter}
\alias{mcpd_filter}
\title{Make or adjust filter using MCPD terminology}
\usage{
mcpd_filter(filter = list(), DOI = NULL, ORIGCTY = NULL,
  SAMPSTAT = NULL)
}
\arguments{
\item{filter}{Existing filters (or blank list if not provided)}

\item{DOI}{Accession DOI}

\item{ORIGCTY}{Country of origin}

\item{SAMPSTAT}{Biological status of sample}
}
\description{
See FAO/Bioversity Multi-Crop Passport Descriptors.
}
\examples{
 # Filter accessions from Mexico and Slovenia
 mcpd_filter(ORIGCTY = c("MEX", "SVN"))


}
