% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{user_login}
\alias{user_login}
\title{Login to Genesys as a user}
\usage{
user_login()
}
\description{
The authorization URL will open in a browser, ask the user to grant
permissions to R and the verification code must be copy-pasted after
you grant access to the client.
}
\seealso{
\code{\link{setup}}
}
