% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesys.R
\name{.fetch_accessions_page}
\alias{.fetch_accessions_page}
\title{Fetch accession passport data page.}
\usage{
.fetch_accessions_page(
  filters = list(),
  page = 0,
  size = 1000,
  selector = NULL
)
}
\value{
Paged data structure
}
\description{
Fetch accession passport data page.
}
\keyword{internal}
