% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{user_login}
\alias{user_login}
\title{Login to Genesys as a user}
\usage{
user_login(redirect_uri = "http://127.0.0.1:48913")
}
\arguments{
\item{redirect_uri}{a custom redirect_uri to submit as part of the authentication request.
This is most useful if the default port is blocked and you wish to specify
another port: redirect_uri = "http://127.0.0.1:44211"
Note that "http://127.0.0.1" is requred and that "http://localhost" will not work.}
}
\description{
The authorization URL will open in a browser, ask the user to grant
permissions to R. After successful authentication the browser will 
display a message:
}
\details{
```
Authentication complete. Please close this page and return to R.
```

Close the browser and return to R.
}
\seealso{
\code{\link{setup}}
}
